+++ src/core/h2.js
@@ -22,7 +22,7 @@ const {
 } = require('http2');
 const { Readable } = require('stream');
 
-const debug = require('debug')('helix-fetch:h2');
+const debug = require('debug')('adobe/fetch:h2');
 
 const { RequestAbortedError } = require('./errors');
 const { decodeStream } = require('../common/utils');
@@ -273,7 +273,7 @@ const request = async (ctx, url, options) => {
     req.once('error', (err) => {
       // error occured during the request
       session.off('error', onSessionError);
-      /* istanbul ignore else */
+      /* istanbul ignore next */
       if (signal) {
         signal.removeEventListener('abort', onAbortSignal);
       }