+++ src/core/request.js
@@ -17,7 +17,7 @@ const tls = require('tls');
 const { types: { isAnyArrayBuffer } } = require('util');
 
 const LRU = require('lru-cache');
-const debug = require('debug')('helix-fetch:core');
+const debug = require('debug')('adobe/fetch:core');
 
 const { RequestAbortedError } = require('./errors');
 const h1 = require('./h1');
@@ -38,7 +38,7 @@ const ALPN_CACHE_SIZE = 100; // # of entries
 const ALPN_CACHE_TTL = 60 * 60 * 1000; // (ms): 1h
 const ALPN_PROTOCOLS = [ALPN_HTTP2, ALPN_HTTP1_1, ALPN_HTTP1_0];
 
-const DEFAULT_USER_AGENT = `helix-fetch/${version}`;
+const DEFAULT_USER_AGENT = `adobe-fetch/${version}`;
 
 // request option defaults
 const DEFAULT_OPTIONS = {
@@ -143,6 +143,14 @@ const determineProtocol = async (ctx, url, signal) => {
 
     default:
       throw new TypeError(`unsupported protocol: ${url.protocol}`);
+  }
+
+  if (ctx.alpnProtocols.length === 1
+    && (ctx.alpnProtocols[0] === ALPN_HTTP1_1 || ctx.alpnProtocols[0] === ALPN_HTTP1_0)) {
+    // shortcut: forced HTTP/1.X, default to HTTP/1.1 (no need to use ALPN to negotiate protocol)
+    protocol = ALPN_HTTP1_1;
+    ctx.alpnCache.set(origin, protocol);
+    return { protocol };
   }
 
   // negotioate via ALPN