+++ src/fetch/index.js
@@ -15,7 +15,7 @@
 const { EventEmitter } = require('events');
 const { Readable } = require('stream');
 
-const debug = require('debug')('helix-fetch');
+const debug = require('debug')('adobe/fetch');
 const LRU = require('lru-cache');
 
 const { Body } = require('./body');
@@ -404,7 +404,7 @@ class FetchContext {
       AbortController,
       AbortSignal,
 
-      // extensions
+      // non-spec extensions
 
       FetchBaseError,
       FetchError,