+++ test/fetch/index.test.js
@@ -210,9 +210,7 @@ testParams.forEach((params) => {
       assert(body.destroyed);
     });
 
-    it('timeoutSignal works (slow response)', async function test() {
-      this.timeout(5000);
-
+    it('timeoutSignal works (slow response)', async () => {
       const signal = timeoutSignal(500);
 
       const method = 'POST';
@@ -232,9 +230,7 @@ testParams.forEach((params) => {
       assert(body.destroyed);
     });
 
-    it('AbortController works (slow response)', async function test() {
-      this.timeout(5000);
-
+    it('AbortController works (slow response)', async () => {
       const controller = new AbortController();
       setTimeout(() => controller.abort(), 1000);
       const { signal } = controller;
@@ -272,9 +268,7 @@ testParams.forEach((params) => {
       }
     });
 
-    it('AbortController works (dripping response)', async function test() {
-      this.timeout(5000);
-
+    it('AbortController works (dripping response)', async () => {
       const FETCH_TIMEOUT = 1000; // ms
       const DRIPPING_DURATION = 2; // seconds
       // doesn't support POST method
@@ -599,6 +593,17 @@ testParams.forEach((params) => {
       const { form: reqForm, headers } = jsonResponseBody;
       assert(headers['Content-Type'].startsWith('multipart/form-data; boundary='));
       assert.deepStrictEqual(reqForm, searchParams);
+    });
+
+    it('returns Set-Cookie headers', async () => {
+      const resp = await fetch(`${baseUrl}/cookies/set?a=1&b=2`, { redirect: 'manual' });
+      // Response headers:
+      // set-cookie: a=1; [Secure; ]Path=/
+      // set-cookie: b=2; [Secure; ]Path=/
+      assert.strictEqual(resp.status, 302);
+      assert(/a=1; (Secure; )?Path=\/, b=2; (Secure; )?Path=\//.test(resp.headers.get('set-cookie')));
+      assert(/a=1; (Secure; )?Path=\//.test(resp.headers.raw()['set-cookie'][0]));
+      assert(/b=2; (Secure; )?Path=\//.test(resp.headers.raw()['set-cookie'][1]));
     });
 
     if (protocol === 'https') {